Fix memory-safety and concurrency defects from security review
Codecs:
- 8SVX: pass the compressed byte count to the Fibonacci-delta decoder and
  bound writes by destination capacity (heap-buffer-overflow on compressed
  files); use fourcc::from_bytes for raw 4-byte chunk IDs (strlen over-read
  in accept())
- VGM: bounds-check every command operand, skip, and data block; validate
  gzip trailer and EOF-offset fields with 64-bit arithmetic; reject data
  offsets past EOF; guard size-8 underflow in ROM blocks; fix 0xe0 consuming
  8 bytes instead of 4 (command-stream desync)

Concurrency:
- serialize the audio callback, stream destruction/moves, device-data swaps,
  and device switching under a process-wide audio_callback_mutex()
- process each stream's callback slice under its state mutex, with volume/
  mute/processor reads under the data mutex; processor list iterated from a
  shared_ptr snapshot
- make playback status flags atomic (getters use a different mutex than the
  writers); atomic stream-token generator; per-device mutex for
  create_stream paths
- restore finish/loop callback dispatch, resolving the current audio_stream
  by token under the callback mutex so moves/destruction cannot dangle the
  pointer; removal precedes invocation so play() from a finish callback
  re-arms the stream
- duration() now takes the state lock (audio source is state-domain)

Memory safety:
- audio_converter: drop size_t -> unsigned int truncating casts in
  allocation paths (OOB writes for inputs above UINT_MAX)
- visualization ring buffer: atomic elements with release/acquire publishing
- decoders registry: return shared_ptr<const decoders_registry> so a
  concurrent set/done cannot free the registry mid-use

Tests:
- add malformed-input regression tests (truncated VGM operands, lying
  EOF/data offsets, oversized data blocks, compressed 8SVX decode, inflated
  BODY size); fix racy plain-bool flag in device-cleanup test

Full suite passes on plain, ASan, and TSan builds (268/268, zero reports).

+++ src/musac/sdk/audio_converter.cc
@@ -33,7 +33,7 @@ buffer<uint8_t> audio_converter::convert(const audio_spec& src_spec,
     
     // Fast path 1: No conversion needed
     if (!needs_conversion(src_spec, dst_spec)) {
-        buffer<uint8_t> output(static_cast<unsigned int>(src_len));
+        buffer<uint8_t> output(src_len);
         std::memcpy(output.data(), src_data, src_len);
         return output;
     }
@@ -42,7 +42,7 @@ buffer<uint8_t> audio_converter::convert(const audio_spec& src_spec,
     if (src_spec.channels == dst_spec.channels && src_spec.freq == dst_spec.freq) {
         if ((src_spec.format == audio_format::s16le && dst_spec.format == audio_format::s16be) ||
             (src_spec.format == audio_format::s16be && dst_spec.format == audio_format::s16le)) {
-            buffer<uint8_t> output(static_cast<unsigned int>(src_len));
+            buffer<uint8_t> output(src_len);
             std::memcpy(output.data(), src_data, src_len);
             detail::fast_swap16_inplace(output.data(), src_len);
             return output;
@@ -51,7 +51,7 @@ buffer<uint8_t> audio_converter::convert(const audio_spec& src_spec,
             (src_spec.format == audio_format::s32be && dst_spec.format == audio_format::s32le) ||
             (src_spec.format == audio_format::f32le && dst_spec.format == audio_format::f32be) ||
             (src_spec.format == audio_format::f32be && dst_spec.format == audio_format::f32le)) {
-            buffer<uint8_t> output(static_cast<unsigned int>(src_len));
+            buffer<uint8_t> output(src_len);
             std::memcpy(output.data(), src_data, src_len);
             detail::fast_swap32_inplace(output.data(), src_len);
             return output;
@@ -297,7 +297,7 @@ size_t audio_converter::stream_converter::process_chunk(const uint8_t* input, si
     if (input_len > 0) {
         // Ensure buffer is large enough
         if (m_pimpl->input_accumulated + input_len > m_pimpl->input_buffer.size()) {
-            m_pimpl->input_buffer.resize(static_cast<unsigned int>(m_pimpl->input_accumulated + input_len + 4096));
+            m_pimpl->input_buffer.resize(m_pimpl->input_accumulated + input_len + 4096);
         }
         std::memcpy(m_pimpl->input_buffer.data() + m_pimpl->input_accumulated, 
                    input, input_len);
